+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Crash in AppHangs when no threads (#2725)
+- MetricKit stack traces (#2723)
 
 ## 8.2.0